rafactor the order features by making class name and val/var name more clear, encapsulate the mutiple navArgs
