+++ BuyTogether/app/src/main/java/com/chloe/shopshare/order/OrderProductAdapter.kt
@@ -1,0 +1,48 @@
+package com.chloe.shopshare.order
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
+import com.chloe.shopshare.data.Product
+import com.chloe.shopshare.databinding.ItemOrderProductBinding
+
+
+class OrderProductAdapter(private val viewModel: OrderViewModel) : ListAdapter<Product, OrderProductAdapter.ViewHolder>(DiffCallback) {
+
+    class ViewHolder(private var binding: ItemOrderProductBinding):
+            RecyclerView.ViewHolder(binding.root) {
+        fun bind(item: Product,viewModel: OrderViewModel) {
+            binding.item = item
+            binding.viewModel = viewModel
+            binding.executePendingBindings()
+        }
+    }
+
+    companion object DiffCallback : DiffUtil.ItemCallback<Product>() {
+        override fun areItemsTheSame(oldItem: Product, newItem: Product): Boolean {
+            return oldItem === newItem
+        }
+        override fun areContentsTheSame(oldItem: Product, newItem: Product): Boolean {
+            return oldItem == newItem
+        }
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
+        return ViewHolder(
+            ItemOrderProductBinding.inflate(
+                        LayoutInflater.from(parent.context), parent, false
+                )
+        )
+    }
+
+    /**
+     * Replaces the contents of a view (invoked by the layout manager)
+     */
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
+        val item = getItem(position)
+        holder.bind(item,viewModel)
+    }
+
+}
