+++ BuyTogether/app/src/main/res/layout/activity_main.xml
@@ -90,7 +90,7 @@
             android:layout_height="@dimen/height_bottom_nav_view_main"
             android:background="@color/white"
             android:elevation="@dimen/elevation_all"
-            android:visibility="@{(viewModel.currentFragmentType == CurrentFragmentType.CHAT_ROOM||viewModel.currentFragmentType == CurrentFragmentType.PARTICIPATE||viewModel.currentFragmentType == CurrentFragmentType.NOTIFY||viewModel.currentFragmentType == CurrentFragmentType.MY_REQUEST||viewModel.currentFragmentType == CurrentFragmentType.ORDER_DETAIL||viewModel.currentFragmentType == CurrentFragmentType.MY_ORDER||viewModel.currentFragmentType == CurrentFragmentType.REQUEST_DETAIL||viewModel.currentFragmentType == CurrentFragmentType.REQUEST||viewModel.currentFragmentType == CurrentFragmentType.LOGIN||viewModel.currentFragmentType == CurrentFragmentType.DETAIL||viewModel.currentFragmentType == CurrentFragmentType.HOST||viewModel.currentFragmentType == CurrentFragmentType.SHOP||viewModel.currentFragmentType == CurrentFragmentType.MANAGE)? View.GONE : View.VISIBLE}"
+            android:visibility="@{(viewModel.currentFragmentType == CurrentFragmentType.CHAT_ROOM||viewModel.currentFragmentType == CurrentFragmentType.ORDER||viewModel.currentFragmentType == CurrentFragmentType.NOTIFY||viewModel.currentFragmentType == CurrentFragmentType.MY_REQUEST||viewModel.currentFragmentType == CurrentFragmentType.ORDER_DETAIL||viewModel.currentFragmentType == CurrentFragmentType.MY_ORDER||viewModel.currentFragmentType == CurrentFragmentType.REQUEST_DETAIL||viewModel.currentFragmentType == CurrentFragmentType.REQUEST||viewModel.currentFragmentType == CurrentFragmentType.LOGIN||viewModel.currentFragmentType == CurrentFragmentType.DETAIL||viewModel.currentFragmentType == CurrentFragmentType.HOST||viewModel.currentFragmentType == CurrentFragmentType.SHOP||viewModel.currentFragmentType == CurrentFragmentType.MANAGE)? View.GONE : View.VISIBLE}"
             app:itemIconTint="@drawable/bottom_nav_item_selector"
             app:labelVisibilityMode="unlabeled"
             app:layout_constraintBottom_toBottomOf="parent"
