+++ BuyTogether/app/src/main/res/layout/fragment_request_detail.xml
@@ -32,7 +32,7 @@
 
 
                 <androidx.constraintlayout.widget.ConstraintLayout
-                    android:id="@+id/detail_img_block"
+                    android:id="@+id/image_block"
                     android:layout_width="0dp"
                     android:layout_height="wrap_content"
                     android:background="#FFFFFF"
@@ -42,7 +42,7 @@
 
 
                     <androidx.recyclerview.widget.RecyclerView
-                        android:id="@+id/recycler_detail_image"
+                        android:id="@+id/recycler_image"
                         android:layout_width="0dp"
                         android:layout_height="0dp"
                         android:orientation="horizontal"
@@ -56,7 +56,7 @@
                         tools:listitem="@layout/item_detail_image"/>
 
                     <androidx.recyclerview.widget.RecyclerView
-                        android:id="@+id/recycler_detail_circles"
+                        android:id="@+id/recycler_image_circles"
                         android:layout_width="0dp"
                         android:layout_height="6dp"
                         android:layout_marginStart="16dp"
@@ -66,7 +66,7 @@
                         app:addDecoration="@{viewModel.decoration}"
                         app:count="@{viewModel.request.image.size}"
                         app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-                        app:layout_constraintBottom_toBottomOf="@+id/recycler_detail_image"
+                        app:layout_constraintBottom_toBottomOf="@+id/recycler_image"
                         app:layout_constraintEnd_toEndOf="parent"
                         app:layout_constraintStart_toStartOf="parent" />
 
@@ -101,7 +101,7 @@
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintHorizontal_bias="0.0"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toBottomOf="@+id/detail_img_block">
+                    app:layout_constraintTop_toBottomOf="@+id/image_block">
 
                     <ImageView
                         android:id="@+id/host_image"
@@ -208,7 +208,7 @@
                 </androidx.constraintlayout.widget.ConstraintLayout>
 
                 <androidx.constraintlayout.widget.ConstraintLayout
-                    android:id="@+id/detail_info_block"
+                    android:id="@+id/information_block"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:background="#FFFFFF"
@@ -255,7 +255,7 @@
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintHorizontal_bias="1.0"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toBottomOf="@+id/detail_info_block">
+                    app:layout_constraintTop_toBottomOf="@+id/information_block">
 
                     <androidx.constraintlayout.widget.ConstraintLayout
                         android:id="@+id/description_title_block"
