+++ BuyTogether/app/src/main/res/layout/item_detail_delivery.xml
@@ -29,7 +29,7 @@
             tools:text="TextView" />
 
         <View
-            android:id="@+id/line_button"
+            android:id="@+id/separation_line"
             android:layout_width="0dp"
             android:layout_height="1dp"
             android:layout_marginTop="8dp"
